lanner-lcd: Convert plcm_drv to a miscdevice
On newer kernels the hardcoded major number (248) might collide with other
upstream devices.  Convert the driver to a miscdevice that dynamically
chooses a minor in the misc major group (10) so we never have this issue
again.  The other benefit of a miscdevice is that the /dev/plcm_drv node
will be created automatically so we don't need to call mknod.  As such,
tweak the postinst script to only call mknod if we are booting a 3.x kernel.

NGFW-11971